backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-socketio==5.11.3